agent/pyproject.toml
[project]
name = "bleikoya-chat-agent"
version = "0.1.0"
description = "Chat search agent for bleikoya.net"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.80.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
serve = "agent.server:main"
